EPA-339 mTLS

+++ build-and-run-docker-image.sh
@@ -2,11 +2,18 @@
 
 #./mvnw clean install
 
+if [ -z "$1" ]
+then
+ echo "Secret path not specified for Konnektor key-store"
+ return 1
+fi
+
 docker rm epa4all
 docker build --progress=plain --no-cache -t epa4all .
 
 docker run -d --name epa4all \
   -p 8090:8090 \
-  -v /Users/bona/Work/ere.health/secret:/opt/epa4all/secret epa4all
+  -v "$1":/opt/epa4all/secret epa4all
 
+# shellcheck disable=SC2028
 echo "docker image epa4all running. \n Status with: 'docker ps' \n Logs with: 'docker logs epa4all' \n Kill with: 'docker kill epa4all'"
